Extracts GpioBridge to an interface
