+++ src/main/java/ch/eitchnet/beaglebone/GpioException.java
@@ -1,7 +1,7 @@
 package ch.eitchnet.beaglebone;
 
 /**
- * General {@link Exception} for exceptional situations while using the {@link GpioBridge}
+ * General {@link Exception} for exceptional situations while using the {@link GpioBridgeImpl}
  * 
  * @author Robert von Burg &lt;eitch@eitchnet.ch&gt;
  */
